+++ packages/theme_tailor_toolbox/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.0.0
+- Support primary constructors
+- Support analyzer 13.0.0
+
 # 3.1.3
 - Use the same versioning for other theme_tailor packages